First step for sharing hit groups between instances: ParameterizedMeshes
A ParameterizedMesh = BLAS + Hit Group Record params, essentially a
unique object that is used in the scene. These unique objects are then
instanced by instances that refer to the parameterized mesh's HG record.
So there will now be as many HG records as there are parameterized
meshes, which will always be <= # of instances